fix(#10): 오탈자 수정
