+++ Coconut-Backend/src/main/java/com/coconut/stock_app/StockAppApplication.java
@@ -5,7 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 
 @SpringBootApplication
-@EntityScan(basePackages = {"com.coconut.stock_app.entity"})
 public class StockAppApplication {
 
 	public static void main(String[] args) {
