+++ Coconut-Backend/src/main/java/com/coconut/stock_app/entity/on_premise/Transaction.java
@@ -1,40 +1,33 @@
-package com.coconut.stock_app.entity.on_premis;
+package com.coconut.stock_app.entity.on_premise;
 
 import jakarta.persistence.*;
 import lombok.*;
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "profit_loss")
+@Table(name = "transactions")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ProfitLoss {
+public class Transaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long profitLossId;
+    private Long transactionId;
 
-    @Column(nullable = false, length = 20)
-    private String stockCode;
-
-    @Column(nullable = false, length = 100)
-    private String stockName;
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private TransactionsTransactionTypeEnum transactionType;
 
     @Column(nullable = false)
-    private BigDecimal purchasePricePerShare;
+    private BigDecimal amount;
 
     @Column(nullable = false)
-    private BigDecimal salePricePerShare;
+    private LocalDate transactionDate;
 
-    private BigDecimal profitRate;
-
-    @Column(nullable = false)
-    private BigDecimal fee;
-
-    @Column(nullable = false)
-    private int saleQuantity;
+    private String description;
 
     @Column(nullable = false)
     private LocalDateTime createdAt;
@@ -46,3 +39,10 @@ public class ProfitLoss {
     @JoinColumn(name = "account_id", referencedColumnName = "accountId", nullable = false)
     private Account account;
 }
+
+
+enum TransactionsTransactionTypeEnum {
+    DEPOSIT,
+    WITHDRAWAL
+}
+
