+++ Coconut-Backend/src/main/java/com/coconut/stock_app/repository/on_premise/OwnedStockRepository.java
@@ -1,6 +1,6 @@
-package com.coconut.stock_app.repository.on_premis;
+package com.coconut.stock_app.repository.on_premise;
 
-import com.coconut.stock_app.entity.on_premis.OwnedStock;
+import com.coconut.stock_app.entity.on_premise.OwnedStock;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
