+++ Coconut-Backend/src/main/java/com/coconut/stock_app/repository/on_premise/TransactionRepository.java
@@ -1,6 +1,6 @@
-package com.coconut.stock_app.repository.on_premis;
+package com.coconut.stock_app.repository.on_premise;
 
-import com.coconut.stock_app.entity.on_premis.Transaction;
+import com.coconut.stock_app.entity.on_premise.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
